+++ CHANGELOG.md
@@ -1,12 +1,11 @@
 # Changelog
 
 # Todo
-- sqlmath - add sql-function IDATETIMETOUNIXEPOCH().
 - none
 
 # v2024.11.1-beta
+- sqlmath - Add sql-function IDATETIMETOEPOCH().
 - sqlmath - Rename sql-functions IDATEXXXFROMTEXT() TO IDATEXXXFROM().
-- sqlmath - Move sql-functions IDATETIMEFROMTEXT(), IDATETIMETOTEXT() from file sqlmath_external_sqlite.c to sqlmath_base.c.
 - sqlmath - Add sql-functions IDATETIMEFROMTEXT(), IDATETIMETOTEXT().
 
 # v2024.10.23